requirements.txt
streamlit
langchain
langchain-together
pypdf2
faiss-cpu
python-dotenv